requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.1
bertopic==0.16.0
blis==1.3.3
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.16.0
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
ecdsa==0.19.1
elastic-transport==9.2.0
elasticsearch==7.17.9
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi>=0.104.1,<0.115.0
filelock==3.20.0
firebase-admin>=6.7.0,<7.0.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.10
google-api-core>=2.28.1
google-api-python-client>=2.187.0
google-auth>=2.43.0
google-auth-httplib2>=0.2.1
google-generativeai>=0.8.3
googleapis-common-protos>=1.72.0
grpcio>=1.76.0
grpcio-status>=1.75.1
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.2.0
httplib2==0.31.0
httptools==0.7.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.1
llvmlite==0.46.0
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
narwhals==2.12.0
networkx==3.6
nltk==3.9.2
numba==0.63.1
numpy==1.26.4
packaging==25.0
pandas==2.3.3
passlib==1.7.4
Pillow==12.0.0
pillow_heif==1.1.1
plotly==6.5.0
preshed==3.0.12
proto-plus>=1.26.1
protobuf>=6.31.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.5
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv>=1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.2.0
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.5.2
starlette==0.27.0
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer==0.15.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions>=4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.24.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.2
websockets==15.0.1
wrapt==2.0.1
